Fixed all samus errors.

+++ Super Metroidvania 3Million/CrossPlatformDesktopProject/Libraries/Sprite/Player/SamusStates/Sprites/LeftWalkSamusSprite.cs
@@ -40,7 +40,7 @@ namespace CrossPlatformDesktopProject.Libraries.Sprite.Player
 
 		public void Update(GameTime gameTime)
         {
-			timer += gameTime.ElapsedGameTime.TotalMilliseconds;
+			timer += (int) gameTime.ElapsedGameTime.TotalMilliseconds;
 			if (timer > interval)
             {
 				if (currentFrame == 3)
@@ -52,7 +52,7 @@ namespace CrossPlatformDesktopProject.Libraries.Sprite.Player
                 }
 				samus.x -= xChange;
 				samus.space = new Rectangle((int)samus.x, (int)samus.y, 64, 64);
-				timer -= gameTime.ElapsedGameTime.TotalMilliseconds;
+				timer -= (int) gameTime.ElapsedGameTime.TotalMilliseconds;
 			}
 
 		}
